Add message recommending extending start/end time
For the analysis to work, the baseline data must cover the scalar
and variometer data. The option to extend the start/end time should
can be used to ensure that the data is correctly covered. It can
also be used to calculate baselines for longer periods of time if
multiple months will be corrected. Because not all data sets require
baselines from before or after the start time the 0 days option is
available.